#981: move ImageSimilarityScorer class to iped-engine module
